+++ lib/stdio/ferror.c
@@ -23,7 +23,7 @@
 int
 ferror (FILE * stream)
 {
-  int fd = (int) stream;
+  int fd = (long) stream;
   if (fd == -1)
     return -1;
   return 0;
